update to 2.4.36.6
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@9607 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/brcm/linux.v24_2/include/linux/signal.h
@@ -160,8 +160,8 @@ static inline void sigfillset(sigset_t *set)
 	default:
 		memset(set, -1, sizeof(sigset_t));
 		break;
-	case 2: set->sig[1] = -1;
-	case 1:	set->sig[0] = -1;
+	case 2: set->sig[1] = (unsigned long)-1;
+	case 1:	set->sig[0] = (unsigned long)-1;
 		break;
 	}
 }
@@ -204,7 +204,7 @@ static inline void siginitsetinv(sigset_t *set, unsigned long mask)
 	default:
 		memset(&set->sig[1], -1, sizeof(long)*(_NSIG_WORDS-1));
 		break;
-	case 2: set->sig[1] = -1;
+	case 2: set->sig[1] = (unsigned long)-1;
 	case 1: ;
 	}
 }
